requirements.txt
jinja2>=3.1.4
pydantic>=2.8.2
pyyaml>=6.0.2
toml>=0.10.2
tqdm>=4.66.5
openai>=1.50.2
langchain>=0.2.14
langchain-core>=0.2.35
langchain-community>=0.2.11
langchain-openai>=0.1.23
jsonschema>=4.23.0
faiss-cpu>=1.8.0
